refactor: add mint card header to not repeat code
