+++ ui/src/views/mint/form-step/mint-form.tsx
@@ -3,34 +3,15 @@ import { Mint } from '../mint.context';
 import { LogoField } from './fields/logo/logo-field';
 import { AppDescriptionField, AppNameField, EnsDomainField } from './fields';
 import { VerifyNFAField } from './fields/verify-nfa-field';
+import { MintCardHeader } from '../mint-card';
 
 export const FormStep = () => {
   const { prevStep, nextStep } = Stepper.useContext();
   const { appName, appDescription, domain } = Mint.useContext();
 
   return (
-    <Card.Container css={{ width: '424px' }}>
-      <Card.Heading
-        title="NFA Details"
-        leftIcon={
-          <IconButton
-            aria-label="Add"
-            colorScheme="gray"
-            variant="link"
-            icon={<Icon name="back" />}
-            css={{ mr: '$2' }}
-            onClick={prevStep}
-          />
-        }
-        rightIcon={
-          <IconButton
-            aria-label="Add"
-            colorScheme="gray"
-            variant="link"
-            icon={<Icon name="info" />}
-          />
-        }
-      />
+    <Card.Container css={{ width: '$107h' }}>
+      <MintCardHeader title="NFA Details" onClickBack={prevStep} />
       <Card.Body>
         <Grid css={{ rowGap: '$6' }}>
           <Grid css={{ rowGap: '$4' }}>
