+++ ui/src/views/mint/github-step/steps/github-connect/index.ts
@@ -1,0 +1,1 @@
+export * from './github-connect-step';
