+++ ui/src/views/mint/github-step/steps/github-repo-configuration/index.ts
@@ -1,0 +1,1 @@
+export * from './repo-configuration';
