+++ ui/src/views/mint/github-step/steps/github-repo-configuration/repo-configuration.tsx
@@ -1,0 +1,12 @@
+import { Card } from '@/components';
+import { RepoConfigurationBody } from './repo-configuration-body';
+import { RepoConfigurationHeader } from './repo-configuration-header';
+
+export const GithubRepoConfiguration: React.FC = () => {
+  return (
+    <Card.Container css={{ width: '$107h' }}>
+      <RepoConfigurationHeader />
+      <RepoConfigurationBody />
+    </Card.Container>
+  );
+};
