+++ ui/src/views/mint/github-step/steps/github-repository-selection/index.ts
@@ -1,0 +1,1 @@
+export * from './github-repository-selection';
