+++ ui/src/views/mint/github-step/steps/index.ts
@@ -1,0 +1,3 @@
+export * from './github-repository-selection';
+export * from './github-repo-configuration';
+export * from './github-connect';
