+++ ui/src/views/mint/mint-card/index.ts
@@ -1,0 +1,1 @@
+export * from './mint-card';
